# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Academic Student Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentStatus = typeof StudentStatus[keyof typeof StudentStatus];


export const StudentStatus = {
  pending: 'pending',
  active: 'active',
  graduated: 'graduated',
  withdrawn: 'withdrawn',
} as const;

export interface Student {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  programId?: number | null;
  /** @nullable */
  programName?: string | null;
  status: StudentStatus;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}

export interface StudentInput {
  firstName: string;
  lastName: string;
  email: string;
  phone?: string;
  country?: string;
  programId?: number;
  bio?: string;
}

export interface StudentUpdate {
  firstName?: string;
  lastName?: string;
  phone?: string;
  country?: string;
  programId?: number;
  bio?: string;
}

export type StudentStatusUpdateStatus = typeof StudentStatusUpdateStatus[keyof typeof StudentStatusUpdateStatus];


export const StudentStatusUpdateStatus = {
  pending: 'pending',
  active: 'active',
  graduated: 'graduated',
  withdrawn: 'withdrawn',
} as const;

export interface StudentStatusUpdate {
  status: StudentStatusUpdateStatus;
}

export type ProgramLevel = typeof ProgramLevel[keyof typeof ProgramLevel];


export const ProgramLevel = {
  certificate: 'certificate',
  masters: 'masters',
  phd_prep: 'phd_prep',
} as const;

export interface Program {
  id: number;
  name: string;
  description: string;
  level: ProgramLevel;
  durationMonths: number;
  courseCount?: number;
}

export type CourseStatus = typeof CourseStatus[keyof typeof CourseStatus];


export const CourseStatus = {
  active: 'active',
  upcoming: 'upcoming',
  completed: 'completed',
} as const;

export interface Course {
  id: number;
  title: string;
  code: string;
  /** @nullable */
  description?: string | null;
  programId: number;
  /** @nullable */
  programName?: string | null;
  instructorName: string;
  credits?: number;
  status: CourseStatus;
  enrollmentCount?: number;
}

export type CourseInputStatus = typeof CourseInputStatus[keyof typeof CourseInputStatus];


export const CourseInputStatus = {
  active: 'active',
  upcoming: 'upcoming',
  completed: 'completed',
} as const;

export interface CourseInput {
  title: string;
  code: string;
  description?: string;
  programId: number;
  instructorName: string;
  credits?: number;
  status?: CourseInputStatus;
}

export interface Enrollment {
  id: number;
  studentId: number;
  courseId: number;
  /** @nullable */
  courseTitle?: string | null;
  /** @nullable */
  courseCode?: string | null;
  /** @nullable */
  instructorName?: string | null;
  /** @nullable */
  credits?: number | null;
  enrolledAt: string;
}

export interface EnrollmentInput {
  studentId: number;
}

export interface Assignment {
  id: number;
  courseId: number;
  /** @nullable */
  courseTitle?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  dueDate: string;
  maxScore: number;
}

export interface AssignmentInput {
  title: string;
  description?: string;
  dueDate: string;
  maxScore: number;
}

export interface AssignmentWithStatus {
  id: number;
  courseId: number;
  /** @nullable */
  courseTitle?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  dueDate: string;
  maxScore: number;
  submitted: boolean;
  /** @nullable */
  submissionId?: number | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  grade?: string | null;
}

export interface Submission {
  id: number;
  assignmentId: number;
  studentId: number;
  /** @nullable */
  content?: string | null;
  submittedAt: string;
}

export interface SubmissionInput {
  studentId: number;
  content?: string;
}

export interface Grade {
  id: number;
  submissionId: number;
  assignmentId?: number;
  /** @nullable */
  assignmentTitle?: string | null;
  /** @nullable */
  courseTitle?: string | null;
  score: number;
  maxScore?: number;
  letterGrade: string;
  /** @nullable */
  feedback?: string | null;
  gradedAt: string;
}

export interface GradeInput {
  score: number;
  letterGrade: string;
  feedback?: string;
}

export type AnnouncementPriority = typeof AnnouncementPriority[keyof typeof AnnouncementPriority];


export const AnnouncementPriority = {
  normal: 'normal',
  important: 'important',
  urgent: 'urgent',
} as const;

export interface Announcement {
  id: number;
  title: string;
  content: string;
  priority?: AnnouncementPriority;
  createdAt: string;
}

export type AnnouncementInputPriority = typeof AnnouncementInputPriority[keyof typeof AnnouncementInputPriority];


export const AnnouncementInputPriority = {
  normal: 'normal',
  important: 'important',
  urgent: 'urgent',
} as const;

export interface AnnouncementInput {
  title: string;
  content: string;
  priority?: AnnouncementInputPriority;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface ProgramCount {
  program: string;
  count: number;
}

export interface DashboardStats {
  totalStudents: number;
  activeStudents: number;
  totalCourses: number;
  totalPrograms: number;
  pendingSubmissions: number;
  recentEnrollments: number;
  studentsByStatus?: StatusCount[];
  enrollmentsByProgram?: ProgramCount[];
}

export interface StudentDashboard {
  student: Student;
  enrolledCourses: number;
  upcomingAssignments: number;
  completedAssignments?: number;
  recentGrades: Grade[];
  /** @nullable */
  gpa: number | null;
}