# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Academic Student Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}/enroll:
    post:
      operationId: enrollInCourse
      tags: [courses]
      summary: Enroll a student in a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /students/{id}/enrollments:
    get:
      operationId: getStudentEnrollments
      tags: [courses]
      summary: Get all course enrollments for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of enrollments with courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"

  # --- Assignments ---
  /courses/{id}/assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments for a course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create an assignment (admin/instructor)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignmentInput"
      responses:
        "201":
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get assignment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /students/{id}/assignments:
    get:
      operationId: getStudentAssignments
      tags: [assignments]
      summary: Get all assignments across enrolled courses for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignments with submission status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignmentWithStatus"
  /assignments/{id}/submit:
    post:
      operationId: submitAssignment
      tags: [assignments]
      summary: Submit an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # --- Grades ---
...
  /dashboard/student/{id}:
    get:
      operationId: getStudentDashboard
      tags: [dashboard]
      summary: Personal dashboard summary for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
components:
  schemas:
...
    Student:
      type: object
      required: [id, firstName, lastName, email, status, createdAt]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        programId:
          type: ["integer", "null"]
        programName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, active, graduated, withdrawn]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Course:
      type: object
      required: [id, title, code, programId, instructorName, status]
      properties:
        id:
          type: integer
        title:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        programId:
          type: integer
        programName:
          type: ["string", "null"]
        instructorName:
          type: string
        credits:
          type: integer
        status:
          type: string
          enum: [active, upcoming, completed]
        enrollmentCount:
          type: integer
    CourseInput:
      type: object
      required: [title, code, programId, instructorName]
      properties:
        title:
          type: string
        code:
          type: string
        description:
          type: string
        programId:
          type: integer
        instructorName:
          type: string
        credits:
          type: integer
        status:
          type: string
          enum: [active, upcoming, completed]

    # --- Enrollment ---
    Enrollment:
      type: object
      required: [id, studentId, courseId, enrolledAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        courseTitle:
          type: ["string", "null"]
        courseCode:
          type: ["string", "null"]
        instructorName:
          type: ["string", "null"]
        credits:
          type: ["integer", "null"]
        enrolledAt:
          type: string
    EnrollmentInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer

    # --- Assignment ---
    Assignment:
      type: object
      required: [id, courseId, title, dueDate, maxScore]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        courseTitle:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: string
        maxScore:
          type: integer
    AssignmentInput:
      type: object
      required: [title, dueDate, maxScore]
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        maxScore:
          type: integer
    AssignmentWithStatus:
      type: object
      required: [id, courseId, title, dueDate, maxScore, submitted]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        courseTitle:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: string
        maxScore:
          type: integer
        submitted:
          type: boolean
        submissionId:
          type: ["integer", "null"]
        score:
          type: ["integer", "null"]
        grade:
          type: ["string", "null"]

    # --- Submission ---
    Submission:
      type: object
      required: [id, assignmentId, studentId, submittedAt]
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        studentId:
          type: integer
        content:
          type: ["string", "null"]
        submittedAt:
          type: string
    SubmissionInput:
      type: object
      required: [studentId]
      properties:
        studentId:
          type: integer
        content:
          type: string

    # --- Grade ---
    Grade:
      type: object
      required: [id, submissionId, score, letterGrade, gradedAt]
      properties:
        id:
          type: integer
        submissionId:
          type: integer
        assignmentId:
          type: integer
        assignmentTitle:
          type: ["string", "null"]
        courseTitle:
          type: ["string", "null"]
        score:
          type: integer
        maxScore:
          type: integer
        letterGrade:
          type: string
        feedback:
          type: ["string", "null"]
        gradedAt:
          type: string
...
    StudentDashboard:
      type: object
      required: [student, enrolledCourses, upcomingAssignments, recentGrades, gpa]
      properties:
        student:
          $ref: "#/components/schemas/Student"
        enrolledCourses:
          type: integer
        upcomingAssignments:
          type: integer
        completedAssignments:
          type: integer
        recentGrades:
          type: array
          items:
            $ref: "#/components/schemas/Grade"
        gpa:
          type: ["number", "null"]
...