# lib/api-zod/src/generated/types/assignmentWithStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Academic Student Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AssignmentWithStatus {
  id: number;
  courseId: number;
  /** @nullable */
  courseTitle?: string | null;
  title: string;
  /** @nullable */
  description?: string | null;
  dueDate: string;
  maxScore: number;
  submitted: boolean;
  /** @nullable */
  submissionId?: number | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  grade?: string | null;
}