# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Academic Student Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentStatus } from './studentStatus';

export interface Student {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  programId?: number | null;
  /** @nullable */
  programName?: string | null;
  status: StudentStatus;
  /** @nullable */
  bio?: string | null;
  createdAt: string;
}